fix(mobile): prevent text leaking through Android glass
